Ejercicio de 10 primeros perfectos

+++ src/pruebas/DiezNumPerfectos.java
@@ -1,19 +1,33 @@
 package pruebas;
 
 /**
- * Calcula los primeros 10 números perfectos.
- * 
+ 
  * Curso 2025-26
  * 
  * @author O.Alvarez
  * @version 1.0
  */
+
+/**
+ * Calcula los primeros 10 números perfectos.
+ * 6
+ * 28
+ * 496
+ * 8128
+ * 33550336
+ * 8589869056
+ * 137438691328
+ * 2305843008139952128
+ * 2658455991569831744654692615953842176
+ * 191561942608236107294793378084303638130997321548169216
+ */
+
 public class DiezNumPerfectos {
 
     public static void main(String[] args) {
 
         int contador = 0;
-        int num = 2; // Empezamos desde 2 (1 nunca es perfecto)
+        long num = 2; // Empezamos desde 2 (1 nunca es perfecto)
 
         System.out.println("Los primeros 10 números perfectos son:");
